before chage db

+++ DuLich/GUI/ManHinhChinh.cs
@@ -45,9 +45,9 @@ namespace DuLich
             {
                 userControl = new ChiTietTouris(tourisAfterUpdate,
                     LoaiDal.GetAll(),
-                    new List<Gia>(),
+                    tourisAfterUpdate.GetListGiaOfTour(),
                     DiaDiemDal.GetAll(),
-                    new List<DiaDiem>(), this, this,this);
+                    tourisAfterUpdate.GetListDiaDiemOfTour(), this, this,this);
                 panel_main_content.Invoke((MethodInvoker)delegate
                 {
                     panel_main_content.Controls.Clear();
@@ -69,7 +69,7 @@ namespace DuLich
                     String message = "";
                     foreach (Doan doan in doanDangKhaiThac)
                     {
-                        message += doan.Name + " /n";
+                        message += "- "+doan.Name + " /n";
                     }
                     DialogResult innerDialogResukt = MessageBox.Show("Tour này đang được sử dụng ở các đoàn : /n" + message + " nên không thể xóa tour này");
                 }
@@ -240,13 +240,13 @@ namespace DuLich
             DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn xóa Giá với mã là :" + gia.MaGia, "", MessageBoxButtons.YesNo);
             if (dialogResult == DialogResult.Yes)
             {
-                List<Doan> doanDangKhaiThac = tourHienTai.GetListDoanOfTour();
+                List<Doan> doanDangKhaiThac = tourHienTai.GetListDoanOfTourWithPrice(gia);
                 if (doanDangKhaiThac.Count > 0)
                 {
                     String message = "";
                     foreach (Doan doan in doanDangKhaiThac)
                     {
-                        message += doan.Name + " \n";
+                        message += "- "+doan.Name + " \n";
                     }
                     DialogResult innerDialogResukt = MessageBox.Show("Giá này đang được áp dụng ở các đoàn : \n" + message + " nên không thể xóa được giá này");
                 }
@@ -304,9 +304,7 @@ namespace DuLich
         private void button1_Click(object sender, EventArgs e)
         {
             userControl = new DanhSachDoan(DoanDal.GetAll(), TourDal.GetAll(),NhanVienDal.GetAll(),KhachDal.GetAll(), this);
-            panel_main_content.SuspendLayout();
             panel_main_content.Controls.Clear();
-            panel_main_content.ResumeLayout();
             panel_main_content.Controls.Add(userControl);
         }
 
@@ -333,15 +331,20 @@ namespace DuLich
 
         public void onDanhSachDoan_XoaClick(Doan doan)
         {
-            doan.Delete().ContinueWith(task =>
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn xóa Đoàn với mã là :" + doan.Id, "", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
             {
-                panel_main_content.Invoke((MethodInvoker)delegate
+                doan.Delete().ContinueWith(task =>
                 {
-                    userControl = new DanhSachKhach(KhachDal.GetAll(), this);
-                    panel_main_content.Controls.Clear();
-                    panel_main_content.Controls.Add(userControl);
+                    panel_main_content.Invoke((MethodInvoker)delegate
+                    {
+                        userControl = new DanhSachDoan(DoanDal.GetAll(), TourDal.GetAll(), NhanVienDal.GetAll(), KhachDal.GetAll(), this);
+                        panel_main_content.Controls.Clear();
+                        panel_main_content.Controls.Add(userControl);
+                    });
                 });
-            });
+            }
+           
         }
     }
 }
@@ -419,7 +422,10 @@ namespace DuLich
 
         public void onDanhSachChiPhi_XoaClick(Doan doanHienTai,ChiPhi chiPhi)
         {
-            doanHienTai.DeleteChiPhi(chiPhi).ContinueWith(task =>
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn xóa chi phí với mã là :" + chiPhi.id, "", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                doanHienTai.DeleteChiPhi(chiPhi).ContinueWith(task =>
             {
                 panel_main_content.Invoke((MethodInvoker)delegate
                 {
@@ -430,6 +436,7 @@ namespace DuLich
                           doanHienTai.GetListKhach()); ;
                 });
             });
+            }
         }
 
         public void onChiTietChiPhi_LuuClick(Doan doanHienTai,ChiPhi chiPhi)
@@ -482,7 +489,10 @@ namespace DuLich
 
         public void onDanhSachPhanCongXoaClick(Doan doanHienTai,PhanCong phanCong)
         {
-            doanHienTai.DeletePhanCong(phanCong).ContinueWith(task =>
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn xóa phân công với mã là :" + phanCong.MaPhanCong, "", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                doanHienTai.DeletePhanCong(phanCong).ContinueWith(task =>
             {
                 panel_main_content.Invoke((MethodInvoker)delegate
                 {
@@ -493,6 +503,7 @@ namespace DuLich
                           doanHienTai.GetListKhach());
                 });
             });
+            }
         }
 
         public void onChiTietPhanCong_LuuClick(Doan doanHienTai,PhanCong phanCong)
@@ -677,7 +688,10 @@ namespace DuLich
 
         public void onDanhSachKhach_XoaClick(Khach khach)
         {
-            khach.Delete().ContinueWith(task =>
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn xóa khách với mã là :" + khach.KhachId, "", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                khach.Delete().ContinueWith(task =>
             {
                 panel_main_content.Invoke((MethodInvoker)delegate
                 {
@@ -686,6 +700,7 @@ namespace DuLich
                     panel_main_content.Controls.Add(userControl);
                 });
             });
+            }
         }
     }
 }
@@ -717,7 +732,10 @@ namespace DuLich
 
     public void onDanhSachNhanVien_XoaClick(NhanVien nhanVien)
     {
-            nhanVien.Delete().ContinueWith(task =>
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc muốn xóa nhân viên với mã là :" + nhanVien.MaNhanVien, "", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                nhanVien.Delete().ContinueWith(task =>
             {
                 panel_main_content.Invoke((MethodInvoker)delegate
                 {
@@ -726,6 +744,7 @@ namespace DuLich
                     panel_main_content.Controls.Add(userControl);
                 });
             });
+            }
         }
     public void onChiTietNhanVien_LuuClick(NhanVien nhanVien)
     {
